requirements.txt
asgiref==3.2.7
Django==3.0.5
gunicorn==20.0.4
Pillow==7.1.1
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise
django-crispy-forms==1.9.0
django-widget-tweaks==1.4.5
django-taggit==1.2.0
